@@ -1,35 +1,20 @@
 source 'https://rubygems.org'
-ruby '2.4.0'
+ruby '3.3.1'
 
 gem 'active_model_serializers'
-gem "actionview", '>= 5.0.7.2'
-gem 'activerecord', '5.0.7.2'
+gem 'coffee-rails'
 gem 'devise'
+gem 'jbuilder'
+gem 'jquery-rails'
+gem 'nokogiri'
 gem 'omniauth'
 gem 'omniauth-google-oauth2'
-gem 'pry'
-
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '5.0.7.2'
-# Use Postgres for launching on heroku
-gem 'pg', '~> 0.21'
-# Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
+gem 'pg'
+gem 'rails', '7.1.3.3'
+gem 'sass-rails'
+# gem 'sdoc', group: :doc
 # gem 'therubyracer', platforms: :ruby
-
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'uglifier'
 
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
@@ -41,14 +26,11 @@ gem 'sdoc', '~> 0.4.0', group: :doc
 # gem 'capistrano-rails', group: :development
 
 group :development, :test do
-  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
+  gem 'pry'
 end
 
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-
-  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'web-console'
   gem 'spring'
 end
