@@ -1,5 +1,5 @@
 class BowlsController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   before_action :set_bowl, only: [:show, :edit, :update, :destroy]
 
   def new
