@@ -1,5 +1,5 @@
 class ScrapsController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   before_action :set_scrap, only: [:show, :edit, :update, :destroy]
   before_action :set_bowl, only: [:new, :edit, :show, :index]
 
