@@ -1,4 +1,4 @@
-class CreateScraps < ActiveRecord::Migration
+class CreateScraps < ActiveRecord::Migration[7.1]
   def change
     create_table :scraps do |t|
       t.integer :user_id
