@@ -21,7 +21,8 @@ module Bowls
     # config.i18n.default_locale = :de
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    # config.active_record.raise_in_transactional_callbacks = true
     config.assets.initialize_on_precompile = false
+    config.load_defaults 7.0
   end
 end
